refactor: Split fucnctions into file #125
